+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 未发布
 
+- `add` 将 `--regex` 参数改为 `--replace`
 - `add` 支持 tar 文件
 - `add` 增加 `--overwrite`
 - HTTP 增加鉴权